cleaned up file structure, memoized some things, removed placeholder game ui img, fixed absolute positioning
