--- front-end/src/domains/game-board/components/background/background.component.tsx
@@ -1,9 +1,0 @@
-import { FC } from "react";
-import "./background.style.scss";
-import TEMP_UI from "./Temp-UI.png";
-
-export const Background: FC = () => (
-  <div className="background">
-    <img className="tempui-img" src={TEMP_UI} alt={"Temp UI For Game Page"}/>
-  </div>
-);
