--- front-end/src/domains/game-board/components/background/background.style.scss
@@ -1,11 +1,0 @@
-.background {
-  background-color: #232323; //PlaceHolder
-  width: 100vw;
-  height: 100vh;
-  overflow: hidden;
-
-  .tempui-img {
-    width: 99%; //These numbers fit on the screen better
-    height: 98%;
-  }
-}
