--- front-end/src/domains/game-board/components/background/index.ts
@@ -1,1 +1,0 @@
-export * from "./background.component.tsx";
