--- front-end/src/domains/game-board/components/game-board/index.ts
@@ -1,1 +1,0 @@
-export * from "./game-hex.component"
