+++ front-end/src/domains/game-board/components/game-grid/index.ts
@@ -1,0 +1,2 @@
+export * from "./game-hex.component";
+export * from "./game-grid.component";
