--- front-end/src/domains/game-board/components/game-placeholder/index.ts
@@ -1,1 +1,0 @@
-export * from "./game-placeholder.component";
