+++ front-end/src/domains/game-board/game-board.component.tsx
@@ -1,10 +1,8 @@
 import { FC } from "react";
-import { Background } from "@domains/game-board/components/background";
-import { Board } from "./components/game-board/game-board.component";
+import { Board } from "./components/game-grid";
 
 export const GameBoard: FC = () => (
   <>
-    <Background />
     <Board />
   </>
 );
