+++ front-end/src/lib/Footer/Footer.style.scss
@@ -7,6 +7,7 @@ footer {
   margin-top: 100px;
   background: #f9f9f9;
   width: 100vw;
+  border: 1px solid red;
 
   .container {
     display: flex;
