+++ front-end/src/views/game-view/Game-View.component.tsx
@@ -3,8 +3,8 @@ import { GameBoard } from "@domains/game-board";
 import { GameUIMenu } from "@domains/game-board/components/game-ui-menu";
 
 export const GameView: FC = () => (
-  <section className="game-view">
+  <>
     <GameBoard />
     <GameUIMenu />
-  </section>
+  </>
 );
